requirements.txt
aiosignal==1.3.1
attrs==23.1.0
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.1.0
coverage==7.2.7
cycler==0.11.0
distlib==0.3.6
exceptiongroup==1.1.2
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
identify==2.5.6
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.17.3
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.7.1
msgpack==1.0.5
nodeenv==1.8.0
numba==0.56.4
numpy==1.23.5
packaging==23.1
pandas==2.0.3
pandas-stubs==2.0.1.230501
patsy==0.5.3
Pillow==10.0.0
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
protobuf==4.23.3
pyarrow==12.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-mock==3.11.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
shap==0.43.0
six==1.16.0
skrebate==0.62
slicer==0.0.7
statsmodels==0.14.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
tune-sklearn==0.4.6
types-pytz==2023.3.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
xgboost==1.7.6
types-toml==0.10.8
Werkzeug==2.2.3
wrapt==1.14.1
pandas-stubs==2.0.1.230501
lime==0.2.0.1
combat==0.3.3
optuna==3.6.1
imbalanced-learn==0.12.2
plotly==5.22.0
kaleido==0.2.1
Levenshtein==0.25.1
BorutaShap==1.0.17
typer==0.17.4
